+++ Tests/MapboxMapsTests/Foundation/MutableRefTests.swift
@@ -27,4 +27,21 @@ class MutableRefTests: XCTestCase {
         XCTAssertEqual(_mutableRef.wrappedValue, "Goodbye, World!")
         XCTAssertEqual(ref.value, "Goodbye, World!")
     }
+
+    func testKeyPathInitialization() {
+        class Foo {
+            var count: Int = 0
+            init(count: Int) {
+                self.count = count
+            }
+        }
+        let instance = Foo(count: 0)
+        var ref = MutableRef(root: instance, keyPath: \.count)
+
+        XCTAssertEqual(ref.wrappedValue, 0)
+        instance.count += 1
+        XCTAssertEqual(ref.wrappedValue, 1)
+        ref.wrappedValue *= 10
+        XCTAssertEqual(instance.count, 10)
+    }
 }